# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  domain.Answers:
    properties:
      answer:
        type: string
      answerid:
        type: integer
    type: object
...
  domain.Events:
    properties:
      applicationclosingdate:
        type: string
      applicationlink:
        type: string
      approved:
        type: boolean
      archived:
        type: boolean
      createdat:
        type: string
      cusatonly:
        type: boolean
      eventdate:
        type: string
      eventid:
        type: integer
      eventpic:
        type: string
      location:
        type: string
      longdiscription:
        type: string
      maxapplications:
        type: integer
      online:
        type: boolean
      organizername:
        type: string
      paid:
        type: boolean
      sex:
        type: string
      shortdiscription:
        type: string
      subevents:
        type: string
      title:
        maxLength: 50
        minLength: 2
        type: string
      websitelink:
        type: string
    required:
    - eventdate
    - eventpic
    - organizername
    - title
    type: object
  domain.Faqas:
    properties:
      answerid:
        type: integer
      createdat:
        type: string
      faqaid:
        type: integer
      organizername:
        type: string
      public:
        type: boolean
      question:
        maxLength: 50
        minLength: 2
        type: string
      title:
        type: string
      username:
        type: string
    required:
    - question
    type: object
  domain.Organizations:
    properties:
      about:
        type: string
      createdat:
        type: string
      createdby:
        type: string
      linkedin:
        type: string
      logo:
        type: string
      organizationid:
        type: integer
      organizationname:
        maxLength: 50
        minLength: 2
        type: string
      verified:
        type: boolean
      websitelink:
        type: string
    required:
    - organizationname
    type: object
  domain.Users:
    properties:
      email:
        type: string
      eventid:
        type: integer
      firstname:
        maxLength: 50
        minLength: 2
        type: string
      lastname:
        maxLength: 50
        minLength: 1
        type: string
      password:
        type: string
      phonenumber:
        type: string
      profile:
        type: string
      userid:
        type: integer
      username:
        maxLength: 50
        minLength: 2
        type: string
      verification:
        type: boolean
      vip:
        type: boolean
    required:
    - email
    - firstname
    - lastname
    - password
    - username
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: sethukumarj.76@gmail.com
    name: API Support
    url: sethukumarj.com
  description: This is an Events Radar project. You can visit the GitHub repository
    at https://github.com/SethukumarJ/Events_Radar_Developement
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin Radar API
  version: "1.0"
paths:
  /accept-invitation:
    get:
      operationId: Accept invitation to join organization
      parameters:
      - description: 'token: '
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Accept invitation to join an organization
      tags:
      - Organization
...
  /organization/admin/add-members:
    post:
      operationId: Add admins for the organizaition
      parameters:
      - description: 'addMembers:'
        in: body
        name: addMembers
        required: true
        schema:
          items:
            type: string
          type: array
      - description: 'OrganizationName: '
        in: query
        name: organizationName
        required: true
        type: string
      - description: member role
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add Admins
      tags:
      - Organization
  /organization/admin/admit-member:
    patch:
      operationId: Admit member
      parameters:
      - description: 'JoinStatusId: '
        in: query
        name: joinstatusid
        required: true
        type: integer
      - description: 'Organization Name :'
        in: query
        name: organizationName
        required: true
        type: string
      - description: member role
        in: query
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Admit member
      tags:
      - Organization
...
  /organization/get-organization:
    get:
      operationId: Get Organizaition by name
      parameters:
      - description: 'OrganizationName: '
        in: query
        name: organizationName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Organization
      tags:
      - Organization
  /organization/join-requests:
    get:
      operationId: Join requests to organization
      parameters:
      - description: 'OrganizationName: '
        in: query
        name: organizationName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Join Requests
      tags:
      - Organization
  /organization/update-event:
    patch:
      operationId: Update event
      parameters:
      - description: event title
        in: query
        name: title
        required: true
        type: string
      - description: 'organizationName: '
        in: query
        name: organizationName
        required: true
        type: string
      - description: update Event with new body
        in: body
        name: UpdateEvent
        required: true
        schema:
          $ref: '#/definitions/domain.Events'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: update event
      tags:
      - Organization
  /user//list-faqas:
    get:
      operationId: list all public faqas
      parameters:
      - description: 'Event title: '
        in: query
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: list all Public faqas
      tags:
      - User
...
  /user/create-organization:
    post:
      operationId: Create Organizatioin from user
      parameters:
      - description: Create organization
        in: body
        name: CreateOrganization
        required: true
        schema:
          $ref: '#/definitions/domain.Organizations'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create Organization
      tags:
      - User
  /user/event/post-question:
    post:
      operationId: User Post Question
      parameters:
      - description: Getting the title of the event
        in: query
        name: title
        required: true
        type: string
      - description: Getting the title of the event
        in: query
        name: organizername
        required: true
        type: string
      - description: Post question
        in: body
        name: PostQuestion
        required: true
        schema:
          $ref: '#/definitions/domain.Faqas'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Post Question function
      tags:
      - User
  /user/event/post/answer:
    post:
      operationId: User Post Answer
      parameters:
      - description: organizationName
        in: query
        name: organizationName
        required: true
        type: string
      - description: Getting the id of the question
        in: query
        name: faqaid
        required: true
        type: string
      - description: Post Answer
        in: body
        name: PostAnswer
        required: true
        schema:
          $ref: '#/definitions/domain.Answers'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Post Answer function
      tags:
      - User
...
  /user/join-organization:
    patch:
      operationId: Join organization
      parameters:
      - description: 'organization name: '
        in: query
        name: organizationName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Joining organization
      tags:
      - User
  /user/list-organizations:
    get:
      operationId: list all registered organizations
      parameters:
      - description: 'Page number: '
        in: query
        name: page
        required: true
        type: integer
      - description: 'Page capacity : '
        in: query
        name: pagesize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: list all registered organizations for user
      tags:
      - User
...
  /user/list/questions:
    get:
      operationId: list all asked questions
      parameters:
      - description: 'Event title: '
        in: query
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: list all Asked questions
      tags:
      - Organization
...
  /user/send-verification:
    post:
      operationId: Send verifiation code via email
      parameters:
      - description: 'Email: '
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: Send verification
      tags:
      - User
...
  /user/update-password:
    patch:
      operationId: Update password
      parameters:
      - description: 'Email: '
        in: query
        name: email
        required: true
        type: string
      - description: update password with new body
        in: body
        name: Updateevent
        required: true
        schema:
          $ref: '#/definitions/domain.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      summary: update password
      tags:
      - User
  /user/update-profile:
    patch:
      operationId: Update userprofile
      parameters:
      - description: update profile with new body
        in: body
        name: UpdateProfile
        required: true
        schema:
          $ref: '#/definitions/domain.Bios'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: update Profileabout
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"